fix: rename src/server.ts so vercel dev doesn't capture it
Vercel's zero-config Node server detection captures a root or src/server.ts
file, expecting a listening server or default fetch handler. Our file
exports a named (req, res) handler wired via api/index.ts + vercel.json
rewrites, so a fresh (unlinked) clone hit 'No fetch handler exported' and
501s. Renaming to src/handler.ts avoids the capture pattern.

Fixes #1104
